Reworked the Organizing specialized implementation
The Organizing specialized implementation now maintains maps / caches
the matched handlers per event data type.
This increases the space complexity, but should provide a small speed up
when there are a lot of different handler types.
Considering how cheap the cast/instanceof check is, it may not be worth
it. The test however does show speed-ups with a big number of event
firings.
So choose wisely!